pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simwise"
version = "0.1.0"
description = "A Python package for satellite simulations"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
keywords = ["simulation", "satellite", "orbit", "aerospace"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy"
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "plotly>=5.0.0",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "pyerfa>=2.0.0",
]

[tool.setuptools]
packages = ["simwise"]